+++ src/backend/joanie/core/views/__init__.py
@@ -1,4 +1,3 @@
 """Views of the ``core`` app of the Joanie project."""
 from .certificate import *
-from .debug import *
 from .redirect import *